refactor: puts shared logic into common file
