+++ src/components/chat/chatScreen/imageChatLogic.js
@@ -17,7 +17,7 @@ import firebase from 'react-native-firebase'
 import {connect} from 'react-redux'
 import ImagePicker from 'react-native-image-crop-picker'
 
-import rsa from '../../rsa'
+import {rsa} from '../../widelySharedLogic'
 import {seenMessages} from '../../../store/actions'
 
 class Chat extends React.Component {
