--- src/components/rsa.js
@@ -1,5 +1,0 @@
-const RSAKey = require('react-native-rsa')
-
-const rsa = new RSAKey()
-
-export default rsa
